backend/app/requirements.txt

fastapi
uvicorn[standard]
requests
python-dotenv
mysql-connector-python
pydantic
pandas
python-multipart
